+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightUI.Skin.cs
@@ -40,6 +40,8 @@ namespace UnityEditor.Rendering.HighDefinition
             public readonly GUIContent colorTemperature = new GUIContent("Temperature", "Specifies a temperature (in Kelvin) HDRP uses to correlate a color for the Light. For reference, White is 6500K.");
             public readonly GUIContent areaLightCookie = new GUIContent("Cookie", "Cookie mask currently assigned to the area light.");
             public readonly GUIContent cookieTextureTypeError = new GUIContent("HDRP does not support the Cookie Texture type, only Default is supported.", EditorGUIUtility.IconContent("console.warnicon").image);
+            public readonly string cookieNonPOT = "HDRP does not support non power of two cookie textures.";
+            public readonly string cookieTooSmall = "Min texture size for cookies is 2x2 pixels.";
 
 
             // Additional light data