+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDBakedReflectionSystem.cs
@@ -176,9 +176,7 @@ namespace UnityEditor.Rendering.HighDefinition
             {
                 // == 4. ==
                 var cubemapSize = (int)hdPipeline.currentPlatformRenderPipelineSettings.lightLoopSettings.reflectionCubemapSize;
-                var planarSize = (int)hdPipeline.currentPlatformRenderPipelineSettings.lightLoopSettings.planarReflectionTextureSize;
                 var cubeRT = HDRenderUtilities.CreateReflectionProbeRenderTarget(cubemapSize);
-                var planarRT = HDRenderUtilities.CreatePlanarProbeRenderTarget(planarSize);
 
                 handle.EnterStage(
                     (int)BakingStages.ReflectionProbes,
@@ -199,13 +197,15 @@ namespace UnityEditor.Rendering.HighDefinition
                     var instanceId = states[index].instanceID;
                     var probe = (HDProbe)EditorUtility.InstanceIDToObject(instanceId);
                     var cacheFile = GetGICacheFileForHDProbe(states[index].probeBakingHash);
+                    var planarRT = HDRenderUtilities.CreatePlanarProbeRenderTarget((int)probe.resolution);
 
                     // Get from cache or render the probe
                     if (!File.Exists(cacheFile))
                         RenderAndWriteToFile(probe, cacheFile, cubeRT, planarRT);
+                    
+                    planarRT.Release();
                 }
                 cubeRT.Release();
-                planarRT.Release();
 
                 // Copy texture from cache
                 for (int i = 0; i < addCount; ++i)
@@ -322,17 +322,17 @@ namespace UnityEditor.Rendering.HighDefinition
             }
 
             var cubemapSize = (int)hdPipeline.currentPlatformRenderPipelineSettings.lightLoopSettings.reflectionCubemapSize;
-            var planarSize = (int)hdPipeline.currentPlatformRenderPipelineSettings.lightLoopSettings.planarReflectionTextureSize;
 
             var cubeRT = HDRenderUtilities.CreateReflectionProbeRenderTarget(cubemapSize);
-            var planarRT = HDRenderUtilities.CreatePlanarProbeRenderTarget(planarSize);
 
             // Render and write the result to disk
             for (int i = 0; i < bakedProbes.Count; ++i)
             {
                 var probe = bakedProbes[i];
                 var bakedTexturePath = HDBakingUtilities.GetBakedTextureFilePath(probe);
+                var planarRT = HDRenderUtilities.CreatePlanarProbeRenderTarget((int)probe.resolution);
                 RenderAndWriteToFile(probe, bakedTexturePath, cubeRT, planarRT);
+                planarRT.Release();
             }
 
             // AssetPipeline bug
@@ -388,7 +388,6 @@ namespace UnityEditor.Rendering.HighDefinition
             }
 
             cubeRT.Release();
-            planarRT.Release();
 
             return true;
         }