+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightDefinition.cs
@@ -43,6 +43,14 @@ namespace UnityEngine.Rendering.HighDefinition
         Refraction
     };
 
+    [GenerateHLSL]
+    public enum CookieMode
+    {
+        None,
+        Clamp,
+        Repeat,
+    }
+
     // These structures share between C# and hlsl need to be align on float4, so we pad them.
     [GenerateHLSL(PackingRules.Exact, false)]
     struct DirectionalLightData
@@ -56,13 +64,15 @@ namespace UnityEngine.Rendering.HighDefinition
         public float   volumetricLightDimmer;   // Replaces 'lightDimer'
 
         public Vector3 forward;
-        public int     cookieIndex;             // -1 if unused (TODO: 16 bit)
+        public CookieMode cookieMode;
+        
+        public Vector4 cookieScaleOffset;
 
         public Vector3 right;                   // Rescaled by (2 / shapeWidth)
-        public int     tileCookie;              // (TODO: use a bitfield)
+        public int     shadowIndex;             // -1 if unused (TODO: 16 bit)
 
         public Vector3 up;                      // Rescaled by (2 / shapeHeight)
-        public int     shadowIndex;             // -1 if unused (TODO: 16 bit)
+        public int     contactShadowIndex;      // -1 if unused (TODO: 16 bit)
 
         public Vector3 color;
         public int     contactShadowMask;       // 0 if unused (TODO: 16 bit)
@@ -93,7 +103,8 @@ namespace UnityEngine.Rendering.HighDefinition
         public float   flareSize;               // Units: radians
 
         public Vector3 surfaceTint;
-        public int     surfaceTextureIndex;     // -1 if unused (TODO: 16 bit)
+
+        public Vector4 surfaceTextureScaleOffset;     // -1 if unused (TODO: 16 bit)
     };
 
     [GenerateHLSL(PackingRules.Exact, false)]
@@ -124,9 +135,11 @@ namespace UnityEngine.Rendering.HighDefinition
         public Vector3 color;
         public float   rangeAttenuationBias;
 
-        public int     cookieIndex;             // -1 if unused
-        public int     tileCookie;              // (TODO: use a bitfield)
+        public CookieMode cookieMode;
+        public int     cookieIndex;             // Texture array index of the point and rectangle light cookies
         public int     shadowIndex;             // -1 if unused (TODO: 16 bit)
+        
+        public Vector4 cookieScaleOffset;       // coordinates of the cookie texture in the atlas
         public int     contactShadowMask;       // negative if unused (TODO: 16 bit)
 
         public Vector3 shadowTint;              // Use to tint shadow color