+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -403,6 +403,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Setting a material's Refraction Model to Thin does not overwrite the Thickness and Transmission Absorption Distance anymore.
 - Removed Wind textures from runtime as wind is no longer built into the pipeline
 - Expose StartSinglePass() and StopSinglePass() as public interface for XRPass
+- Replaced the Texture array for 2D cookies (spot, area and directional lights) and for planar reflections by an atlas.
 
 ## [7.1.1] - 2019-09-05